LayoutTests/inspector/debugger/selected-call-frame-after-formatting-source-expected.txt

Tests selected call frame does not change when pretty-print is toggled.

Bug 70906
Debugger was enabled.
Set timer for test function.
Script execution paused.
Script execution resumed.
Debugger was disabled.
